requirements.txt
pyyaml
flask
mutagen
